Refactor discord_client.py into smaller modules
